Enhancement: Use npm scripts for building documentation.
