+++ gulpfile.js
@@ -7,75 +7,9 @@
 
 'use strict';
 
-const path = require( 'path' );
 const gulp = require( 'gulp' );
 
-// Documentation. -------------------------------------------------------------
-
-gulp.task( 'docs', () => {
-	const skipLiveSnippets = process.argv.includes( '--skip-snippets' );
-	const skipApi = process.argv.includes( '--skip-api' );
-	const production = process.argv.includes( '--production' );
-
-	if ( skipApi ) {
-		const fs = require( 'fs' );
-		const apiJsonPath = './docs/api/output.json';
-
-		if ( fs.existsSync( apiJsonPath ) ) {
-			fs.unlinkSync( apiJsonPath );
-		}
-
-		return runUmberto( {
-			skipLiveSnippets,
-			skipApi,
-			production
-		} );
-	}
-
-	// Simple way to reuse existing api/output.json:
-	// return Promise.resolve()
-	return buildApiDocs()
-		.then( () => {
-			return runUmberto( {
-				skipLiveSnippets,
-				production
-			} );
-		} );
-} );
-
-gulp.task( 'docs:api', buildApiDocs );
-
-function buildApiDocs() {
-	assertIsInstalled( '@ckeditor/ckeditor5-dev-docs' );
-
-	const ckeditor5Docs = require( '@ckeditor/ckeditor5-dev-docs' );
-
-	return ckeditor5Docs
-		.build( {
-			readmePath: path.join( process.cwd(), 'README.md' ),
-			sourceFiles: [
-				process.cwd() + '/packages/ckeditor5-*/src/**/*.@(js|jsdoc)',
-				'!' + process.cwd() + '/packages/ckeditor5-*/src/lib/**/*.js',
-				'!' + process.cwd() + '/packages/ckeditor5-build-*/src/**/*.js'
-			],
-			validateOnly: process.argv[ 3 ] == '--validate-only'
-		} );
-}
-
-function runUmberto( options ) {
-	assertIsInstalled( 'umberto' );
-	const umberto = require( 'umberto' );
-
-	return umberto.buildSingleProject( {
-		configDir: 'docs',
-		clean: true,
-		skipLiveSnippets: options.skipLiveSnippets,
-		snippetOptions: {
-			production: options.production
-		},
-		skipApi: options.skipApi
-	} );
-}
+const assertIsInstalled = require( './scripts/util/assertisinstalled' );
 
 // Translations. --------------------------------------------------------------
 
@@ -118,17 +52,3 @@ gulp.task( 'release:dependencies', () => {
 			packages: 'packages'
 		} );
 } );
-
-// Utils. ---------------------------------------------------------------------
-
-function assertIsInstalled( packageName ) {
-	try {
-		require( packageName + '/package.json' );
-	} catch ( err ) {
-		console.error( `Error: Cannot find package '${ packageName }'.\n` );
-		console.error( 'You need to install optional dependencies.' );
-		console.error( 'Run: \'npm run install-optional-dependencies\'.' );
-
-		process.exit( 1 );
-	}
-}
