+++ package.json
@@ -71,6 +71,8 @@
     "precommit": "lint-staged",
     "test": "node --max_old_space_size=4096 ./node_modules/.bin/ckeditor5-dev-tests",
     "test:manual": "node --max_old_space_size=4096 ./node_modules/.bin/ckeditor5-dev-tests-manual",
+    "docs": "./scripts/docs/build-docs.js",
+    "docs:api": "./scripts/docs/build-api-docs.js",
     "install-optional-dependencies": "./scripts/install-optional-dependencies.sh",
     "switch-to-dev-dev": "./scripts/switch-to-dev-dev.sh",
     "build-and-publish-docs": "./scripts/docs/build-and-publish.js"
